reorganize client project into folders
