--- client/src/components/Chatroom.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Chatroom() {
-  return <span>Chatroom</span>;
-}
-
-export default Chatroom;
