+++ client/src/components/Gallery/FriendList.js
