+++ client/src/components/Login/Form.css
@@ -1,4 +1,4 @@
-@import "../reset.css";
+@import "../../reset.css";
 
 * {
   box-sizing: border-box;
