--- client/src/components/Toolbar.js
@@ -1,1 +1,0 @@
-import React from "react";
