+++ client/src/components/WorkspacePanels/ChatroomPanel/Chatroom.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "../BoxStyle.css";
+
+function Chatroom() {
+  return (
+    <div className="panel">
+      <h1 className="panel__title">Chat</h1>
+    </div>
+  );
+}
+
+export default Chatroom;
