+++ client/src/components/WorkspacePanels/InstrumentsPanel/Instruments.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "../BoxStyle.css";
+
+function Instruments() {
+  return (
+    <div className="panel">
+      <h1 className="panel__title">Instruments</h1>
+    </div>
+  );
+}
+
+export default Instruments;
