+++ client/src/components/WorkspacePanels/MainWorkspacePanel/MainWorkspace.css
@@ -1,0 +1,3 @@
+.panel {
+  /* Main editing window styling */
+}
