+++ client/src/components/WorkspacePanels/MainWorkspacePanel/MainWorkspace.js
@@ -1,0 +1,13 @@
+import React from "react";
+import "../BoxStyle.css";
+import "./MainWorkspace.css";
+
+function MainWorkspace() {
+  return (
+    <div className="panel">
+      <h1 className="panel__title">Main Workspace</h1>
+    </div>
+  );
+}
+
+export default MainWorkspace;
