--- client/src/routes/Entry.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Form from "../components/Form";
-
-function Entry() {
-  return <Form />;
-}
-
-export default Entry;
