+++ client/src/routes/LoginRoute/Login.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Form from "../../components/Login/Form";
+
+function Login() {
+  return <Form />;
+}
+
+export default Login;
