+++ client/src/routes/WorkspaceRoute/Workspace.css
@@ -1,4 +1,4 @@
-@import "../reset.css";
+@import "../../reset.css";
 
 html,
 body {
