+++ client/src/routes/WorkspaceRoute/Workspace.js
@@ -1,17 +1,11 @@
 import React from "react";
-import Chatroom from "../components/Chatroom";
-import Samples from "../components/Samples";
-import Effects from "../components/Effects";
+import Chatroom from "../../components/WorkspacePanels/ChatroomPanel/Chatroom";
+import Samples from "../../components/WorkspacePanels/SamplesPanel/Samples";
+import Effects from "../../components/WorkspacePanels/EffectsPanel/Effects";
+import Instruments from "../../components/WorkspacePanels/InstrumentsPanel/Instruments";
+import MainWorkspace from "../../components/WorkspacePanels/MainWorkspacePanel/MainWorkspace";
 
 import "./Workspace.css";
-
-function Instruments() {
-  return <span>Instruments</span>;
-}
-
-function MainWorkspace() {
-  return <span>Main workspace</span>;
-}
 
 function Workspace() {
   return (
